removed angular and added api project
