+++ SocialMediaApp/API/Controllers/WeatherForecastController.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace WebApp.Controllers
+namespace API.Controllers
 {
     [ApiController]
     [Route("[controller]")]
