+++ SocialMediaApp/SocialMediaApp.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29409.12
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApp", "WebApp\WebApp.csproj", "{DF4A28FA-DE7C-485C-9BBB-1F6894B43035}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Library", "Library\Library.csproj", "{86F3E0BC-ECCF-4F0B-BBF3-3AD263515D5F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Library", "Library\Library.csproj", "{86F3E0BC-ECCF-4F0B-BBF3-3AD263515D5F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataAccess", "DataAccess\DataAccess.csproj", "{FB608FA3-0E31-488F-BA61-BEC5B596AAC6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataAccess", "DataAccess\DataAccess.csproj", "{FB608FA3-0E31-488F-BA61-BEC5B596AAC6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Testing", "Testing\Testing.csproj", "{B9ECBDE7-B52F-409A-98C5-05C1A80A04A7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Testing", "Testing\Testing.csproj", "{B9ECBDE7-B52F-409A-98C5-05C1A80A04A7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API", "API\API.csproj", "{636F24A6-C6C0-4CE5-A7AB-56D7BED55466}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DF4A28FA-DE7C-485C-9BBB-1F6894B43035}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DF4A28FA-DE7C-485C-9BBB-1F6894B43035}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DF4A28FA-DE7C-485C-9BBB-1F6894B43035}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DF4A28FA-DE7C-485C-9BBB-1F6894B43035}.Release|Any CPU.Build.0 = Release|Any CPU
 		{86F3E0BC-ECCF-4F0B-BBF3-3AD263515D5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{86F3E0BC-ECCF-4F0B-BBF3-3AD263515D5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{86F3E0BC-ECCF-4F0B-BBF3-3AD263515D5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -33,6 +29,10 @@ Global
 		{B9ECBDE7-B52F-409A-98C5-05C1A80A04A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B9ECBDE7-B52F-409A-98C5-05C1A80A04A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B9ECBDE7-B52F-409A-98C5-05C1A80A04A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{636F24A6-C6C0-4CE5-A7AB-56D7BED55466}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{636F24A6-C6C0-4CE5-A7AB-56D7BED55466}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{636F24A6-C6C0-4CE5-A7AB-56D7BED55466}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{636F24A6-C6C0-4CE5-A7AB-56D7BED55466}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
