--- SocialMediaApp/WebApp/ClientApp/src/app/app.component.html
@@ -1,6 +1,0 @@
-<body>
-  <app-nav-menu></app-nav-menu>
-  <div class="container">
-    <router-outlet></router-outlet>
-  </div>
-</body>
