--- SocialMediaApp/WebApp/ClientApp/src/app/counter/counter.component.html
@@ -1,7 +1,0 @@
-<h1>Counter</h1>
-
-<p>This is a simple example of an Angular component.</p>
-
-<p aria-live="polite">Current count: <strong>{{ currentCount }}</strong></p>
-
-<button class="btn btn-primary" (click)="incrementCounter()">Increment</button>
