--- SocialMediaApp/WebApp/ClientApp/src/app/home/home.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-home',
-  templateUrl: './home.component.html',
-})
-export class HomeComponent {
-}
