--- SocialMediaApp/WebApp/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using WebApp
-@namespace WebApp.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
