--- SocialMediaApp/WebApp/appsettings.json
@@ -1,8 +1,0 @@
-{
-  "Logging": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    },
-"AllowedHosts": "*"
-}
